Working training for hpctoolkit and programl

+++ compiler2_service/model/rllib_trainer.py
@@ -57,6 +57,8 @@ import importlib
 import time
 
 from ray.tune.integration.wandb import WandbLoggerCallback
+# from ray.air.callbacks.wandb import WandbLoggerCallback
+
 from compiler2_service.paths import COMPILER2_ROOT
 from os.path import exists
 import wandb
@@ -103,7 +105,7 @@ def make_env():
 
 
 class RLlibTrainer:
-    def __init__(self, profiler, trainer, dataset, steps, size, eval_size, network, sweep_count, eval_time, ray_mode='local', wandb_key_path=str(COMPILER2_ROOT) + "/wandb_key.txt") -> None:        
+    def __init__(self, profiler, trainer, dataset, steps, size, eval_size, network, sweep_count, eval_time, ray_mode='local') -> None:        
         self.ray_start(ray_mode)
         self.wandb_dict = {}
 
@@ -115,7 +117,8 @@ class RLlibTrainer:
         self.trainer_name = trainer
         self.profiler = profiler
         self.algorithm, trainer = trainer.split('.') # expected: algorithm.trainer
-        self.trainer = getattr(importlib.import_module(f'ray.rllib.algorithms.{self.algorithm}'), trainer)
+        # self.trainer = getattr(importlib.import_module(f'ray.rllib.algorithms.{self.algorithm}'), trainer)
+        self.trainer = getattr(importlib.import_module(f'ray.rllib.agents.{self.algorithm}'), trainer)
         self.config = importlib.import_module(f"compiler2_service.model.config.{self.algorithm}.{trainer}").get_config(profiler, sweep_count)
         self.dataset = dataset
         self.size = size
@@ -135,7 +138,6 @@ class RLlibTrainer:
         self.my_artifacts_end = self.my_artifacts/'end'
         self.my_artifacts_results = self.my_artifacts/'results'
         self.device = 'cuda' if torch.cuda.is_available() else 'cpu'
-        self.wandb_key_path = wandb_key_path
         self.checkpoint_path = None
         self.policy_model = None
         self.train_benchmarks = []
@@ -153,7 +155,7 @@ class RLlibTrainer:
         # os.mkdir(self.my_artifacts_end)
         self.my_artifacts_start.mkdir(parents=True)
         self.my_artifacts_end.mkdir(parents=True)
-        
+
         ModelCatalog.register_custom_model(
             "my_model", self.network
         )
@@ -194,7 +196,7 @@ class RLlibTrainer:
 
     def ray_start(self, ray_mode):
         if ray.is_initialized(): ray.shutdown()
-        
+        # breakpoint()
         if ray_mode == 'slurm':
             ray_address = os.environ["RAY_ADDRESS"] if "RAY_ADDRESS" in os.environ else "auto"
             head_node_ip = os.environ["HEAD_NODE_IP"] if "HEAD_NODE_IP" in os.environ else "127.0.0.1"
@@ -202,9 +204,12 @@ class RLlibTrainer:
             print('SLURM options: ', ray_address, head_node_ip, redis_password)
             ray.init(address=ray_address, _node_ip_address=head_node_ip, _redis_password=redis_password)    
         elif ray_mode == 'local':
-            ray.init(local_mode=True, ignore_reinit_error=True)
+            # ray.init(local_mode=True, ignore_reinit_error=True)
+            # ray.init(local_mode=True, num_cpus=8, num_gpus=1,  _temp_dir=f"/scratch/dx4/tmp", object_store_memory=78643200)
+            ray.init(local_mode=True, _temp_dir=f"/scratch/dx4/tmp")
         elif ray_mode == 'non-local':
-            ray.init(local_mode=False, num_gpus=1, object_store_memory=100 * 1024 * 1024, ignore_reinit_error=True)
+            # ray.init(local_mode=False, num_cpus=os.cpu_count(), num_gpus=torch.cuda.device_count(), object_store_memory=100 * 1024 * 1024, ignore_reinit_error=True)
+            ray.init(local_mode=False, num_cpus=65, num_gpus=1, object_store_memory=100 * 1024 * 1024, ignore_reinit_error=True)
         else:
             print('Ray mode must be: slurm, local, non-local')
 
@@ -254,7 +259,7 @@ class RLlibTrainer:
         if self.checkpoint_start != None:
             train_iter += self.checkpoint_start['training_iteration']
 
-        breakpoint()
+        # breakpoint()
         analysis = tune.run(
             self.trainer,
             config=self.config, 
@@ -264,15 +269,13 @@ class RLlibTrainer:
             # checkpoint_freq=10,
             checkpoint_at_end=True,
             num_samples=max(1, sweep_count),
-            stop={'training_iteration': train_iter, 'episode_reward_mean': stop_reward, 'timesteps_total': 1},   
+            stop={'training_iteration': train_iter, 'episode_reward_mean': stop_reward}, #, 'timesteps_total': 1},   
             fail_fast=True,
             reuse_actors=True,
             callbacks=[
-                # MyCallback(self),
                 WandbLoggerCallback(
                     project=self.wandb_project_name,
-                    api_key_file=self.wandb_key_path,
-                    log_config=False,
+                    log_config=True,
                 )
             ],
         )
@@ -281,21 +284,22 @@ class RLlibTrainer:
         if len(analysis.trials):
             trial = analysis.trials[0]
         else: 
+            print(f"DEJAN: len(analysis.trials) <= 0 !, it is {analysis}")
             breakpoint()
 
-        return self.get_agent(trial_config=trial.config, checkpoint_path_str=str(trial.checkpoint.dir_or_data), trial_id=trial.trial_id) # .value -> .dir_or_data for ray 2.1
+        return self.get_agent(trial_config=trial.config, checkpoint_path_str=str(trial.checkpoint.value), trial_id=trial.trial_id) # .value -> .dir_or_data for ray 2.1
 
 
     def get_agent(self, trial_config, checkpoint_path_str, trial_id=None):
                 
         trial_config["explore"] = False
-        self.agent = self.trainer.get_default_config() \
-                        .from_dict(trial_config) \
-                        .build(env='compiler_gym')
-        # self.agent = self.trainer( 
-        #     env="compiler_gym",
-        #     config=trial_config
-        # )
+        # self.agent = self.trainer.get_default_config() \
+        #                 .from_dict(trial_config) \
+        #                 .build(env='compiler_gym')
+        self.agent = self.trainer( 
+            env="compiler_gym",
+            config=trial_config
+        )
 
         self.agent.restore(checkpoint_path_str)
 
@@ -317,8 +321,12 @@ class RLlibTrainer:
         policy_path = self.my_artifacts_end/trial_id#/'policy_model.pt'
         os.makedirs(policy_path)#.parent)
         my_artifacts_checkpoint_dir = self.my_artifacts_end/trial_id/checkpoint_path.name # parent.
-        shutil.copytree(checkpoint_path, my_artifacts_checkpoint_dir)
-        with open(my_artifacts_checkpoint_dir/'config.json', "w") as f: json.dump(config, f)
+
+        if checkpoint_path.is_file():
+            shutil.copyfile(checkpoint_path, my_artifacts_checkpoint_dir)
+        else:
+            shutil.copytree(checkpoint_path, my_artifacts_checkpoint_dir)
+            with open(my_artifacts_checkpoint_dir/'config.json', "w") as f: json.dump(config, f)
 
         self.send_to_wandb(wandb_run_id=f'{self.wandb_project_url}/{trial_id}', wandb_dict=self.wandb_dict, path=self.my_artifacts_end/trial_id)
 
@@ -360,7 +368,7 @@ class RLlibTrainer:
         self.wandb_dict[f'final_performance_train'] = self.evaluator.evaluate(env=self.env, agent=agent, benchmarks=self.wandb_dict['eval_train_benchmarks'], profiler=self.profiler, csv_file=f'{self.my_artifacts_results}/{self.profiler}_train.csv')
         self.wandb_dict[f'final_performance_test'] = self.evaluator.evaluate(env=self.env, agent=agent, benchmarks=self.wandb_dict['eval_test_benchmarks'], profiler=self.profiler, csv_file=f'{self.my_artifacts_results}/{self.profiler}_test.csv')
         
-        breakpoint()
+        # breakpoint()
         if self.trial_id:
             self.send_to_wandb(wandb_run_id=f'{self.wandb_project_url}/{self.trial_id}', wandb_dict=self.wandb_dict, path=self.my_artifacts_results)
